Fix layout of maplibre controls by only using pico in explicit parts of the layout.
picocss is conflicting with maplibre - most notably the maplibre
controls have a silly amount of padding, so the controls look wonky.

Instead of applying picocss everywhere, we now "allow list" subtrees of
the DOM with the "pico" class.

It's kind of annoying, but it kind of seems like it has to be this way -
if we're going to have two strong design systems (pico and maplibre),
they need to be confined to their own respective domains.