update to latest (not yet released) geo
NOTE: the test fixture changes appear to be due to https://github.com/georust/geo/pull/1310,
which reversed the Polygon winding on the output from geo::boolean_operations.

Now the exterior is wound counter-clockwise and the interior clockwise.

This output convention matches what's used elsewhere in geo and also in
the geojson spec.

In practice many tools are tolerant of both, but it's probably better to
follow convention here.