It's not a "shortcut" unless it goes through the interior.
Originally we talked about making this more sophisticated - looking for
including in "editable_roads" but I think *regardless* of if we're
letting the user edit the perimeter, we wouldn't include trips solely
along the perimeter as shortcuts.

== Perf

This is primarily a correctness fix, but it also substantially speeds up the
calculation of shortcuts. It's still 2x slower vs. "intersection" (pre
turn restriction) routing, but 2x is a lot better than 10x!

shortcuts in bristol_east
                        time:   [2.0087 ms 2.0097 ms 2.0108 ms]
                        change: [-81.521% -81.424% -81.332%] (p = 0.00 < 0.05)
                        Performance has improved.
shortcuts in bristol_west
                        time:   [5.7247 ms 5.7953 ms 5.8633 ms]
                        change: [-77.296% -76.972% -76.653%] (p = 0.00 < 0.05)
                        Performance has improved.
shortcuts in strasbourg time:   [24.363 ms 25.531 ms 26.718 ms]
                        change: [-79.893% -78.657% -77.258%] (p = 0.00 < 0.05)
                        Performance has improved.