Prioritize neighborhood by SIMD (UI only) (#152)
* Expose "neighbourhood" boundary to rust.

Previously we dealt with these as "opaque" GeoJSON Features, but for the
prioritization work, we're going to be working directly with these
entities in rust, so it makes sense to have a 1st class rust
representation, rather than going through GeoJSON.

The GeoJSON representation is now only used for serialization, and
relegated to the js/rust boundary.

* Show area on hover.

* Mock SIMD output

* WIP: new binary to gen simd data

* WIP: show prioritization in left column

Also zhuzh the project list a bit

TODO

-[ ] Only cnt
-[/] Color map accordingly
 -[x] simd is colored
 -[ ] area is colored strangely
-[x] info hierarchy of neighborhood list

* Highlight list item when hovering map.

Unfortunately, the reverse seems tricky: highlighting the map when
hovering on the list:

I had a go at using `<JoinedData>` to associate a highlighted
feature-state. Doing so requires joining the source layer by layer id
and thus specifying a specific layer id for the GeoJSON.

But then the JoinedData calls some maplibre code that complained:

> GeoJSON sources cannot have a sourceLayer parameter.

* wip: url params

* extract selection widget

* Prioritize on autogen

* better name?

* unpatch now that PR is merged

* minimize diff, clarify code placement

* revert failed ui experiment

* Only prioritize in CNT

I'm still showing "area" when hovering over neighbourhoods in the
PickeNeighbourhoodMode and AutoBoundariesMode, even in non-CNT, cuz why
not?

* more exciting font icons