prep data for simd prioritization, and tie it into the app.
The current methodology uses area proportional accumulation, but we
might revisit.

Runtime to prepare the data is about 4 minutes on my machine, about 2/3 of which is spent
computing the tricky topology of the Highlands and Islands region.

> Time since start 244.382083083s

There is some low hanging fruit to speed up the spatial joins in wasm, but I left that as TODO
for now. I'll come back to that next week.