Fix bug loading any saved project using Overpass. #170
A week or so ago, we switched from polygons to multipolygons for study
area boundaries, to handle island regions in Scotland for CNT. But for
projects using Overpass imports, we need to handle either polygons or
multipolygons.

Fix itself is in https://github.com/a-b-street/svelte-utils/commit/11d15bbab588acf14506b26882a813d980da59ce.