Revert buffering single selection
We solved the "sliver" problem by buffering the selected boundary.

Although it "looks nicer" to have the buffer be consistent when toggling
between 1 and 2 selected neighbourhoods, buffering affects the
aggregated prioritization metrics.

But when you have only one neighbourhood selected, the "aggregated"
metrics should match that of the one boundary on the map **exactly**.

Having the visual buffer jitter a little between 1 and 2 selections is
the lesser evil.