make buildings more prominent
Buildings are useful context for orienting yourself on the map,
especially when you're a local expert who recognizes the landmarks.

With the colored cells, it's a bit hard to see the buildings. The 3-d
building layer which we used to show while zoomed in was a actually a
little easier to distinguish. Now that we're showing the 2-d layer, I've
compensated by making it a bit darker.