Combine choose/draw neighbourhood UI
I had set out to work on "one neighbourhood per project", with the idea
being that, after creating a project, users should immediately be dumped
into the "create neighbourhood" screen.

The problem is, we have two ways to create neighbourhoods: by picking
auto gen boundaries and drawing them by hand.

This change combines the 2 UI modes. Originally it was in pursuit of the
"one neighbourhood per project", but this change streamlines the UI for
everyone I think.

Now the user doesn't have to pick in advance (which could be confusing
if you don't already know the difference).

Plus, now you can see prioritization metrics for your "hand drawn"
neighbourhoods, which was missing before.