PERF: skip trivial intersection calculation
The `within` check seems to give 2% boost, and the majority is from the
`contains` check. Those this would vary based on the size of the
severance boundaries vs. the population zones. In practice they are
usually correlated: denser areas means smaller population zones, but it
also big roads are closer together in the city than they are in rural
areas.

In any case, we were already computing the de-9im matrix, so these
additional checks are free.

Change vs before this commit:

inverness/generate auto boundaries (and stats)
                        time:   [20.322 s 20.568 s 20.809 s]
                        change: [-38.620% -37.101% -35.783%] (p = 0.00 < 0.05)