Cleared "one way" button (#269)
* Avoid subtle border rasterization artifacts

At least on Safari, there's an undesirable small (~1px) white border
on the exterior of the red circle visible when the "turn restrictions"
control is enabled (overlaying a non-white background).

This is not visible when using the svg icon.

* Clarify "travel flow" tool icon.

This was sourced from https://www.gov.uk/guidance/traffic-sign-images#terms-and-conditions

> You may reproduce traffic-sign images free of charge and without having to
> seek permission, but you must reproduce them accurately and not in a
> misleading context (eg not on roadside billboards where they could mislead
> drivers). You should also include a statement that these images are Crown
> copyright.

For now I've put a statement in the About modal, though I understand
that component might be revamped soon.

* cleared tooltip copy: "one-way"

Even though it's a bit of a lie, (it's also a "two-way" tool), "one-way"
might be a more useful label for people to intuit the purpose of this
tool.

* Typos

---------

Co-authored-by: Dustin Carlino <dabreegster@gmail.com>