"pickable neighbourhood" doesn't look disabled
On the pick neighbourhood screen currently, the area is colored gray,
which communicates "disabled". E.g. we use the same gray color as a mask
to "blot out" the rest of the world outside of the study area.

Now we use the same boundary outline style as used when building the
neighbourhood boundary.

I think it's nice to throw back to the same design we use elsewhere to
communiate neighbourhood boundaries.

I also think it's nice to match the "highlight" color on the map with
what we show in the sidebar when hovering.

I also like that the lack of mask when *not* hoevered means you can see
the interior of the neighbourhoods.

The major downside to this approach is that without a visible "fill" it
might not be obvious enough that the neighbourhoods represent a 2-d
area.

I could easily be convinced to add a low-opacity fill to the unhovered
state. It shouldn't be gray though. Maybe blue like we show for the LAD
outlines on the CNT title page?

Or maybe it's fine as phrased.