Fix occasionally missing hover state.
There seems to be a bug in svelte-maplibre (or maplibre?). Presumably
the issue is related to quickly switching between hovered features.

To reproduce pan into and out of one of the CNT boundaries. About 10% of
the time, the background will not be colored appropriately for the hover
state. Interestingly, the popup will still appear.

I cannot reproduce the issue after removing "hover" handling from the
linestring borders. We don't need it anywhere.