Less obtrusive "Border Entry" arrow
Previously, we drew an arrow on the rust size, as a map feature (sized in
meters). It's hard to get the sizing right that way, because they are
either too small when zoomed out or too large when zoomed in.

Using a _symbol_ layer instead of a _feature_ lets us decouple the size
of the rendered symbol from the zoom level.

Another problem this attempts to solve: "long ago" before "main roads"
we classified "perimeter roads" along the perimeter of the neighbourhood
boundary. This meant that the border entry arrows were drawn over top
the neighbourhood mask, so it was clear what they were referencing.

However, with the switch from "Perimeter" road to "Main" road, it means
you have to account for these designs on a potentially interior main
road. The previous offset now means that these border arrows were drawn
over part of the  neighbourhood, and it was sometimes unclear what they
were referencing.

We address that by positioning the border entry "on top" of the main road.
There is a downside to this - the intersections are more crowded with
symbols now, especially if you have turn restrictions. On balance, I
think it's a good tradeoff.