Link to "predict impact" mode from neighbourhood.
One shortcut I took is opting out of tracking the "true" back state for
the DebugDemandMode, accessible via PredictImpactDetailMode. We just
always go back to pick-neighbouhood. I don't think people will get very
confused by this in practice.

If we want to continue to simulate a stack properly so that hitting "back" 3 times
returns you to the same page from DebugDemandMode, we'd need to track
`prevPrevPrevMode` which seems a bit much.

At some point it might make sense to track an explicit stack of previous
`modes`, rather than trying to manually track the "prevMode" and
"prevPrevMode".